+++ src/main/java/oxff/top/ui/AboutPanel.java
@@ -12,7 +12,7 @@ import java.net.URISyntaxException;
  */
 public class AboutPanel extends JPanel {
 
-    private static final String VERSION = "2.16.5";
+    private static final String VERSION = "2.17.0";
     private static final String AUTHOR = "githubnull";
     private static final String LICENSE = "Apache License 2.0";
     private static final String GITHUB_URL = "https://github.com/githubnull/repeater-manager";